# projeto-rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de controle de pacientes. Os doutores fazem login (OAuth2 password + JWT). Os dados do paciente (cadastro, consulta, alteração, remoção) ficam no serviço SOAP (projeto-soap); este REST guarda os documentos gerados em cada vinda e as finalizações de plantão. Cada vinda do paciente ao hospital é uma passagem (uma por dia) com os seus documentos: a admissão abre a passagem em aguardando_triagem, a triagem leva para aguardando_consulta e a consulta dá alta. O resumo do paciente usa só as passagens anteriores. Quadro em tempo real em GET / (WebSocket em /ws?token=...).",
        "title": "Projeto REST API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o doutor dono do token enviado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Usuário autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Cria um novo doutor (usuário) que poderá fazer login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar doutor",
                "parameters": [
                    {
                        "description": "Nome e senha",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "description": "Fluxo OAuth2 \"password\": valida nome/senha e devolve um JWT Bearer. Aceita JSON {nome, senha} ou form-urlencoded (grant_type=password\u0026username=...\u0026password=...).",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Obter token (login)",
                "parameters": [
                    {
                        "description": "Credenciais",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.OAuthError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.OAuthError"
                        }
                    }
                }
            }
        },
        "/finalizacao": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tudo o que o doutor autenticado registrou (chegadas, triagens, consultas, correções) desde a última finalização. É o que ele revisa e comenta antes de finalizar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finalização"
                ],
                "summary": "Resumo para finalizar o plantão",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PreviaFinalizacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grava a finalização com o comentário geral e os comentários opcionais de cada movimentação, e encerra a sessão do doutor: o token atual deixa de valer e o quadro é desconectado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finalização"
                ],
                "summary": "Finalizar o plantão",
                "parameters": [
                    {
                        "description": "Comentários do doutor",
                        "name": "finalizacao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FinalizarDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Finalizacao"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/finalizacoes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Plantões já finalizados pelo doutor autenticado, do mais recente para o mais antigo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finalização"
                ],
                "summary": "Minhas finalizações",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Finalizacao"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ComentarioMovimentacao": {
            "type": "object",
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "movimentacao_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.FinalizarDTO": {
            "type": "object",
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "comentarios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ComentarioMovimentacao"
                    }
                }
            }
        },
        "dto.LoginDTO": {
            "type": "object",
            "required": [
                "nome",
                "senha"
            ],
            "properties": {
                "nome": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
        "dto.OAuthError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_grant"
                },
                "error_description": {
                    "type": "string"
                }
            }
        },
...
        "dto.PreviaFinalizacao": {
            "type": "object",
            "properties": {
                "fim": {
                    "type": "string"
                },
                "inicio": {
                    "type": "string"
                },
                "movimentacoes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movimentacao"
                    }
                },
                "pacientes": {
                    "type": "integer"
                },
                "totais": {
                    "description": "Quantas movimentações de cada tipo (chegada, triagem, consulta...)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "dto.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
        "dto.UserDTO": {
            "type": "object",
            "required": [
                "nome",
                "senha"
            ],
            "properties": {
                "nome": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
...
        "model.Finalizacao": {
            "type": "object",
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "doutor_id": {
                    "type": "integer"
                },
                "fim": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inicio": {
                    "type": "string"
                },
                "movimentacoes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movimentacao"
                    }
                }
            }
        },
        "model.Movimentacao": {
            "type": "object",
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "doutor_id": {
                    "type": "integer"
                },
                "em": {
                    "type": "string"
                },
                "finalizacao_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "paciente_id": {
                    "type": "integer"
                },
                "paciente_nome": {
                    "type": "string"
                },
                "passagem_id": {
                    "type": "integer"
                },
                "tipo": {
                    "type": "string",
                    "example": "triagem"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Cole o access_token obtido em POST /auth/token (com ou sem o prefixo \"Bearer \")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}